requirements.txt

torch
numpy
gym
playwright
streamlit
Pillow
imageio
matplotlib
tensorzero @ git+https://github.com/tensorzero/tensorzero.git
